Moved uploadDummyFile function to FileActions trait
