+++ it/uk/gov/hmrc/fileupload/support/FileActions.scala
@@ -1,5 +1,6 @@
 package uk.gov.hmrc.fileupload.support
 
+import org.scalatest.time.{Seconds, Span}
 import play.api.Play.current
 import play.api.libs.ws.{WS, WSResponse}
 import uk.gov.hmrc.fileupload.{EnvelopeId, FileId}
@@ -31,4 +32,14 @@ trait FileActions extends ActionsSupport {
       .url(s"$url/envelopes/$envelopeId/files/$fileId/metadata")
       .get()
       .futureValue
+
+  def uploadDummyFile(envelopeId: EnvelopeId, fileId: FileId): WSResponse = {
+    WS.url(s"$url/upload/envelopes/$envelopeId/files/$fileId")
+      .withHeaders("Content-Type" -> "multipart/form-data; boundary=---011000010111000001101001",
+        "X-Request-ID" -> "someId",
+        "X-Session-ID" -> "someId",
+        "X-Requested-With" -> "someId")
+      .post("-----011000010111000001101001\r\nContent-Disposition: form-data; name=\"file1\"; filename=\"test.txt\"\r\nContent-Type: text/plain\r\n\r\nsomeTextContents\r\n-----011000010111000001101001--")
+      .futureValue(PatienceConfig(timeout = Span(10, Seconds)))
+  }
 }
